Primera paralelización con OpenCL - preprocesamiento[A

+++ opencl/yangCL.c
@@ -180,7 +180,7 @@ void lcs_opencl(char *a, char *b, int m, int n, int block_count, int thread_coun
     checkError(ret, "Setting kernel arguments");
     ret = clSetKernelArg(kernel, 2, sizeof (cl_mem), (void *)&d_alfabeto);
     checkError(ret, "Setting kernel arguments");
-    ret = clSetKernelArg(kernel, 3, sizeof(long), &n);
+    ret = clSetKernelArg(kernel, 3, sizeof(int), &n);
     checkError(ret, "Setting kernel arguments");
     ret = clSetKernelArg(kernel, 4, sizeof(int), &l);
     checkError(ret, "Setting kernel arguments");
